Test.Tasty.Runners.AntXML

Description

Run a TestTree and produce an XML file summarising the test results in the same schema that would be produced by Apache Ant's JUnit test runner. This schema can be intepreted by the Jenkins continuous integration server, amongst other tools.

Documentation

antXMLRunner :: Ingredient

To run tests using this ingredient, use defaultMainWithIngredients, passing antXMLRunner as one possible ingredient. This ingredient will run tests if you pass the --xml command line option. For example, --xml=junit.xml will run all the tests and generate junit.xml as output.

newtype AntXMLPath

Constructors

AntXMLPath FilePath